Vectorized occlusion based map computation
