+++ xaitk_saliency/impls/saliency/vis_sal_similarity/sliding_window.py
@@ -49,7 +49,7 @@ class SlidingWindow(PerturbImage):
 
         mask_num = len(rows) * len(cols)
         masks = np.ones((mask_num, image_size[0], image_size[1]),
-                        dtype=np.float64)
+                        dtype=np.uint8)
         i = 0
         for r in rows:
             for c in cols:
@@ -79,42 +79,9 @@ class SlidingWindow(PerturbImage):
         self,
         ref_image: PIL.Image.Image
     ) -> Tuple[List[PIL.Image.Image], np.ndarray]:
-        """
-        Transform an input reference image into a number of perturbed
-        variations along with mask matrices indicating the perturbed regions.
 
-        Output images should have the same shape as the input reference image,
-        including channels.
-
-        Output mask matrices should be congruent in length to the number of
-        perturbed images output, as well as share the same height and width
-        dimensions.
-        These masks should indicate the regions in the corresponding perturbed
-        image that has been modified.
-        Values should be in the [0, 1] range, where a value closer to 1.0
-        indicate areas of the image that are *unperturbed*.
-        Note that output mask matrices *may be* of a floating-point type in
-        order to allow for fractional perturbation.
-
-        The resolution of masking is controlled by the window_size and stride
-        of siding window.
-
-        :param ref_image:
-            Reference image to generate perturbations from.
-        :return: Tuple of perturbed images and the window masks detailing perturbation
-            areas.
-        """
         img_dims = ref_image.size
         self.masks = self.generate_block_masks(self.window_size,
                                                self.stride, img_dims)
         masked_images = generate_masked_images(self.masks, ref_image)
         return masked_images, self.masks
-
-    def __call__(
-        self,
-        ref_image: PIL.Image.Image
-    ) -> Tuple[List[PIL.Image.Image], np.ndarray]:
-        """
-        Alias for :meth:`.PerturbImage.perturb`.
-        """
-        return self.perturb(ref_image)
